# specs/001-interview-evidence-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Interview Evidence Platform API
  version: 1.0.0
  description: |
    Contract baseline for the Korean-language pilot. All company routes enforce the company_id
    from the authenticated principal. Applicant routes enforce the invitation-scoped session
    cookie; neither API accepts company_id from an untrusted request body.
servers:
  - url: /v1
...
paths:
  /me:
    get:
      operationId: getCurrentCompanyUser
      x-owner-lane: A
      tags: [Company]
      security:
        - CompanyBearer: []
      responses:
        "200":
          description: Authenticated company user and tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUserView"
        default:
          $ref: "#/components/responses/Error"
  /positions:
    get:
      operationId: listPositions
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Tenant-scoped positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createPosition
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionCreate"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        default:
          $ref: "#/components/responses/Error"
  /positions/{position_id}:
    get:
      operationId: getPosition
      x-owner-lane: A
      tags:
        - Hiring
      security:
        - CompanyBearer: []
      parameters:
        - $ref: '#/components/parameters/PositionId'
      responses:
        '200':
          description: Tenant-scoped position details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
        default:
          $ref: '#/components/responses/Error'
    patch:
      operationId: updatePosition
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
        - $ref: "#/components/parameters/IfMatchVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionUpdate"
      responses:
        "200":
          description: Position settings and lifecycle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        default:
          $ref: "#/components/responses/Error"
  /interviewer-profiles:
    get:
      operationId: listInterviewerProfiles
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Tenant-scoped reusable AI interviewer profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewerProfilePage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createInterviewerProfile
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewerProfileCreate"
      responses:
        "201":
          description: Reusable AI interviewer profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewerProfile"
        default:
          $ref: "#/components/responses/Error"
  /positions/{position_id}/competency-model-versions:
    get:
      operationId: listCompetencyModelVersions
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Position criterion versions, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetencyModelVersionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createCompetencyModelVersion
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetencyModelVersionCreate"
      responses:
        "201":
          description: Immutable draft version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetencyModelVersion"
        default:
          $ref: "#/components/responses/Error"
  /competency-model-versions/{version_id}/publish:
    post:
      operationId: publishCompetencyModelVersion
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchVersion"
      responses:
        "200":
          description: Published immutable criterion version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetencyModelVersion"
        default:
          $ref: "#/components/responses/Error"
  /positions/{position_id}/invitations:
    get:
      operationId: listInvitations
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Invitations and progress projections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createInvitations
      x-owner-lane: A
      tags: [Hiring]
      security:
        - CompanyBearer: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationBatchCreate"
      responses:
        "202":
          description: Invitations accepted for issuance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationBatchResult"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    CompanyBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApplicantSession:
      type: apiKey
      in: cookie
      name: iep_applicant_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    PositionId:
      name: position_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Structured error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    CompanyUserView:
      type: object
      additionalProperties: false
      required: [company_user_id, company_id, email, status]
      properties:
        company_user_id: {type: string, format: uuid}
        company_id: {type: string, format: uuid}
        email: {type: string, format: email}
        status: {type: string, enum: [invited, active, disabled]}
    PositionCreate:
      type: object
      additionalProperties: false
      required: [title, description]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        description: {type: string, minLength: 1, maxLength: 20000}
        role_type: {type: [string, "null"], maxLength: 100}
        headcount: {type: [integer, "null"], minimum: 1, maximum: 10000}
        recruitment_start_at: {type: [string, "null"], format: date}
        recruitment_end_at: {type: [string, "null"], format: date}
    PositionUpdate:
      type: object
      additionalProperties: false
      required: [title, description, status]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        description: {type: string, minLength: 1, maxLength: 20000}
        role_type: {type: [string, "null"], maxLength: 100}
        headcount: {type: [integer, "null"], minimum: 1, maximum: 10000}
        recruitment_start_at: {type: [string, "null"], format: date}
        recruitment_end_at: {type: [string, "null"], format: date}
        status: {type: string, enum: [draft, active, closed]}
    Position:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/PositionCreate"
        - type: object
          required: [position_id, status, row_version, created_at]
          properties:
            position_id: {type: string, format: uuid}
            status: {type: string, enum: [draft, active, closed]}
            row_version: {type: integer, minimum: 1}
            created_at: {type: string, format: date-time}
    PositionPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/Position"}
        next_cursor: {type: [string, "null"]}
    InterviewerProfileCreate:
      type: object
      additionalProperties: false
      required: [name, tone, voice_id]
      properties:
        name: {type: string, minLength: 1, maxLength: 80}
        tone:
          type: string
          enum: [calm, friendly, analytical, concise]
        voice_id: {type: string, minLength: 1, maxLength: 100}
...
    InterviewerProfilePage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/InterviewerProfile"}
        next_cursor: {type: [string, "null"]}
...